requirements.txt
annotated-types==0.7.0
attmap==0.13.2
attrs==24.2.0
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
cffi==1.17.1
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.7
eido==0.2.3
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
fsspec==2024.9.0
gmpy2 @ file:///C:/ci_311/gmpy2_1677743390134/work
gpt4all==2.8.2
huggingface-hub==0.25.0
humanfriendly==10.0
idna==3.10
Jinja2 @ file:///C:/b/abs_92fccttino/croot/jinja2_1716993447201/work
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
logmuse==0.2.8
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/b/abs_f55mv94vyg/croot/mkl_fft_1725370278455/work
mkl_random @ file:///C:/b/abs_21ydbzdu8d/croot/mkl_random_1725370276095/work
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
nltk==3.9.1
numpy @ file:///C:/b/abs_0123vcxhf8/croot/numpy_and_numpy_base_1725470331966/work/dist/numpy-2.0.1-cp311-cp311-win_amd64.whl#sha256=46eb5850478cda64bda6d2f2d3d6ad8ceeb7cc168b83532e4f9dbbcb42e22ab2
oyaml==1.0
packaging==24.1
pandas==2.2.2
pephubclient==0.4.4
peppy==0.40.6
pillow @ file:///C:/b/abs_32o8er3uqp/croot/pillow_1721059447598/work
piper==0.14.2
pipestat==0.10.2
playsound==1.3.0
protobuf==5.28.1
psutil==6.0.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pypiwin32==223
pyreadline3==3.5.3
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-dateutil==2.9.0.post0
pyttsx3==2.97
pytz==2024.2
pywin32==306
PyYAML @ file:///C:/b/abs_782o3mbw7z/croot/pyyaml_1698096085010/work
referencing==0.35.1
regex==2024.9.11
requests @ file:///C:/b/abs_9frifg92q2/croot/requests_1721410901096/work
rich==13.8.1
rpds-py==0.20.0
safetensors==0.4.5
sentencepiece==0.2.0
shellingham==1.5.4
simpleaudio==1.0.4
six==1.16.0
soundfile==0.12.1
SpeechRecognition==3.10.4
sympy @ file:///C:/b/abs_4e4p71hdj_/croot/sympy_1724938208509/work
tokenizers==0.19.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
typer==0.12.5
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata==2024.1
ubiquerg==0.8.0
urllib3 @ file:///C:/b/abs_a7hvzm4y95/croot/urllib3_1718912661242/work
whisper==1.1.10
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
yacman==0.9.3
